# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: |
    API specification for Legal Lawyer AI backend.

    Safe Mode:
    - Only endpoints that declare the `X-Safe-Mode` header enforce it.
    - When `X-Safe-Mode: 1`, those endpoints block writes and return 403.
servers:
  - url: https://gowaned-beckham-unlawful.ngrok-free.dev
  - url: http://127.0.0.1:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    safeModeHeader:
      type: apiKey
      in: header
      name: X-Safe-Mode
      description: "Set to '1' to enable safe mode (read-only, no persistence)"
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
    RateLimitError:
      type: object
      properties:
        message: { type: string, example: "Rate limit exceeded" }
        error: { type: string, example: "Too Many Requests" }
    OkIdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
...
    UserPublicSchema:
      type: object
      required:
        [
          id,
          name,
          email,
          phone,
          cnic,
          fatherName,
          fatherCnic,
          motherName,
          motherCnic,
          city,
          gender,
          age,
          totalSiblings,
          brothers,
          sisters,
          avatarPath,
          timezone,
          language,
          isAdmin,
          isEmailVerified,
        ]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: "Ahmed Khan" }
        email: { type: string, format: email, example: "ahmed@example.com" }
        phone: { type: string, example: "03001234567" }
        cnic: { type: string, example: "12345-1234567-1" }
        fatherName: { type: string, nullable: true }
        fatherCnic: { type: string, nullable: true }
        motherName: { type: string, nullable: true }
        motherCnic: { type: string, nullable: true }
        city: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        age: { type: integer, nullable: true }
        avatarPath: { type: string, nullable: true }
        timezone: { type: string, example: "Asia/Karachi" }
        language: { type: string, enum: [en, ur], example: "en" }
        totalSiblings: { type: integer }
        brothers: { type: integer }
        sisters: { type: integer }
        isAdmin: { type: boolean }
        isEmailVerified: { type: boolean }
...
    ActivityEvent:
      type: object
      required: [type, payload, createdAt]
      properties:
        type: { type: string }
        payload: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
...
    BookmarkSchema:
      type: object
      required: [id, itemType, itemId]
      properties:
        id: { type: integer, format: int64 }
        itemType: { type: string, enum: [right, template, pathway] }
        itemId: { type: integer, format: int64 }

    # Draft Schemas
...
    UnauthorizedErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Invalid token" }
        error: { type: string, example: "Unauthorized" }
...
security: [] # default none; per-route add as needed
paths:
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      description: Returns complete user profile for authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPublicSchema" }
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
    put:
      tags: [Users]
      summary: Update current user profile
      description: |
        Update user profile information.

        **Rate Limit:** 30 requests per minute
        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                phone: { type: string, minLength: 5, maxLength: 30 }
                cnic: { type: string, minLength: 5, maxLength: 30 }
                fatherName: { type: string, nullable: true }
                fatherCnic: { type: string, nullable: true }
                motherName: { type: string, nullable: true }
                motherCnic: { type: string, nullable: true }
                city: { type: string, nullable: true }
                gender: { type: string, nullable: true }
                age: { type: integer, nullable: true }
                totalSiblings: { type: integer }
                brothers: { type: integer }
                sisters: { type: integer }
                timezone: { type: string }
                language: { type: string, enum: [en, ur] }
            example:
              name: Ahmed Updated
              phone: "03009876543"
              city: Karachi
              age: 29
              language: en
      responses:
        "200":
          description: Profile updated successfully
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Invalid language
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema:
                type: object
                required: [ok, reason]
                properties:
                  ok: { type: boolean, example: false }
                  reason: { type: string, example: "Safe mode" }
        "429":
          description: Too Many Requests (30 per minute)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/users/me/avatar:
    post:
      tags: [Users]
      summary: Upload user avatar
      description: |
        Upload profile picture.

        **Rate Limit:** 10 requests per minute
        **Constraints:** Enforced by server configuration
        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Avatar file"
      responses:
        "200":
          description: Avatar uploaded successfully
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: object
                required: [avatarPath]
                properties:
                  avatarPath:
                    { type: string, example: "avatars/abc123_avatar.jpg" }
        "400":
          description: Invalid file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema:
                type: object
                required: [ok, reason]
                properties:
                  ok: { type: boolean, example: false }
                  reason: { type: string, example: "Safe mode" }
        "429":
          description: Too Many Requests (10 per minute)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/users/me/bookmarks:
    post:
      tags: [Users]
      summary: Add bookmark
      description: |
        Bookmark a content item (right, template, or pathway).

        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemType, itemId]
              properties:
                itemType: { type: string, enum: [right, template, pathway] }
                itemId: { type: integer, format: int64 }
            example:
              itemType: right
              itemId: 1
      responses:
        "201":
          description: Bookmark created successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkIdResponse" }
        "400":
          description: Invalid itemType
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema:
                type: object
                required: [ok, reason]
                properties:
                  ok: { type: boolean, example: false }
                  reason: { type: string, example: "Safe mode" }
    get:
      tags: [Users]
      summary: List user bookmarks
      description: Returns all bookmarks for authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Bookmarks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BookmarkSchema" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/users/me/bookmarks/{bookmarkId}:
    delete:
      tags: [Users]
      summary: Delete bookmark
      description: |
        Remove a bookmark.

        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      parameters:
        - in: path
          name: bookmarkId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Bookmark deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema:
                type: object
                required: [ok, reason]
                properties:
                  ok: { type: boolean, example: false }
                  reason: { type: string, example: "Safe mode" }
  /api/v1/users/me/activity:
    post:
      tags: [Users]
      summary: Log user activity event
      description: |
        Log analytics event (e.g., viewed_right, searched, created_draft).

        **Safe Mode:** Silently ignored when `X-Safe-Mode: 1` (returns success without logging)
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventType: { type: string, maxLength: 30 }
                payload: { type: object, additionalProperties: true }
            example:
              eventType: viewed_right
              payload:
                rightId: 5
                timestamp: "2025-01-15T10:30:00Z"
      responses:
        "200":
          description: Activity logged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
    get:
      tags: [Users]
      summary: Get user activity history
      description: Returns last 6 activity events for user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Activity history retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityEvent" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
    delete:
      tags: [Users]
      summary: Clear user activity history
      description: Deletes all activity events for the user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Activity history cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }

  # -------------------- EMERGENCY CONTACTS --------------------
...